% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement_plot.R
\name{plot_rearrangements}
\alias{plot_rearrangements}
\title{Plot rearrangements from SV and CNV data}
\usage{
plot_rearrangements(
  bedpe,
  chrs,
  cn_bedgraph = NULL,
  segments = NULL,
  yrange = NULL,
  ideogram = T,
  cn_cex = 0.5,
  lwd = 0.75,
  cn_win_size = 1e+05,
  BFB_ids = c(),
  arrow_ln = 0.15,
  xlim = NULL,
  chr_lim = NULL,
  annot = NULL,
  main = NULL,
  CN_SV_gap = F,
  ascat_tbl = NULL,
  specific_SV_cols = NA,
  p_ylim_quantile = 0.999,
  yaxis_ticks = NA,
  y_sv1 = 0.3,
  y_sv2 = 0.75,
  yaxis_side = 4
)
}
\arguments{
\item{bedpe}{File handle for BEDPE containing SVs}

\item{chrs}{The chromosomes to plot for}

\item{cn_bedgraph}{File handle for bg file contain copy number information. Defaults to NULL.}

\item{specific_SV_cols}{(default: NA): if set, then SVs will be plotted with the colors provided in the vector (one entry for each SV)}

\item{CN_SV_gap:}{(default: False) If set to true then a gap of size yrange_size will be left between copy number (CN) and structural variant (SV) data.}

\item{y_sv1:}{yrange\link{2}*(1+y_sv1) is where the lower SVs are plotted}

\item{y_sv2:}{yrange\link{2}*(1+y_sv2) is where the higher SVs are plotted}

\item{yaxis_side:}{side where to plot y axis. Takes same values as side argument in mtext (1=bottom, 2=left, 3=top, 4=right)}
}
\description{
This function plots structural variant and CNV data against the genome..
}
\examples{
fn_bedpe <- '/data/CGP_donor_A/sanger/A_SV_4_0_12/A_vs_B.annot.bedpe'
fn_cn <- '/data/CGP_donor_A/sanger/A_SV_4_0_12/intermediates/B.ngscn.abs_cn.bg.gz'
bedpe <- read.table(fn_bedpe, header = F, sep = "\t", stringsAsFactors = F)
zz <- gzfile('file.csv.gz','rt')  
cn    <- read.table(zz,    header = F, sep = "\t", stringsAsFactors = F, colClasses = c("factor", "numeric", "numeric", "numeric"))
this_sample <- 'test'
fpath_out <- '~/dev/testplot'
pdf(paste0(fpath_out,".pdf"), h = 4, w = 10)
plot_rearrangements(bedpe = bedpe, chrs=c(1), cn_bedgraph = cn, cn_win_size = 1e4, cn_cex = 0.3)
dev.off()
}
\keyword{rearrangment}
